Added Red Highlighting
People often post scams on KSL cars to separate other people from their
money without anything in return.  One sign of this is a listing without
any phone number.  Added red row highlighting for listings that do not
have a phone number in the right spot. The listing could have a phone
number in the description, so this is just a suggestion (as the tooltip
on the checkbox column implies).